+++ packages/editor/src/components/collaborators-presence/index.tsx
@@ -110,8 +110,7 @@ export function CollaboratorsPresence( {
 										collaboratorState.collaboratorInfo.name
 									}
 									borderColor={ getAvatarBorderColor(
-										collaboratorState.collaboratorInfo.id ??
-											collaboratorState.clientId
+										collaboratorState.clientId
 									) }
 									size="small"
 								/>